% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_package_sources.R
\name{compile_package_sources}
\alias{compile_package_sources}
\title{Compile Package Sources into a Single Markdown File}
\usage{
compile_package_sources(
  package_path,
  output_file = "combined_package_sources.md"
)
}
\arguments{
\item{package_path}{A character string indicating the path to the local R package directory.}

\item{output_file}{A character string for the name (and optional path) of the output \code{.md} file.}
}
\value{
The function writes a single \code{.md} file (specified by \code{output_file}) containing
the concatenated contents of the README, R scripts, Rd files, and C/C++ source files.
}
\description{
This function reads and concatenates an R package's \code{README.md}, all \code{.R} scripts in the \verb{R/} folder,
all \code{.Rd} files in the \verb{man/} folder, and all recognized source files in \verb{src/} (e.g., \code{.cpp}, \code{.h}, \code{.win}).
The result is a single \code{.md} file containing all these concatenated texts.
}
\examples{
\dontrun{
# Usage example:
compile_package_sources(
  package_path = "path/to/your/R-package",
  output_file  = "combined_package_sources.md"
)
}

}
