% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhist.R
\name{myhist}
\alias{myhist}
\title{Plot a Histogram with Optional Normal Distribution Overlay}
\usage{
myhist(data, var_name, normal_dens = TRUE)
}
\arguments{
\item{data}{A data frame containing the variable to be plotted.}

\item{var_name}{The name of the variable to plot (unquoted).}

\item{normal_dens}{Logical. If \code{TRUE}, overlays a normal density curve. Defaults to \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
This function creates a histogram of the specified variable. Optionally, a normal
distribution curve can be overlaid, along with a vertical line indicating the mean.
}
\details{
\itemize{
\item The histogram displays the density of the variable.
\item If \code{normal_dens = TRUE}, a blue normal distribution curve overlays the histogram.
\item A red dashed vertical line marks the mean of the variable.
\item The plot uses the \code{theme_lucid()} from the \strong{see} package for styling.
}
}
\examples{
\dontrun{
library(ggplot2)
library(see)
myhist(mtcars, mpg)
myhist(mtcars, mpg, normal_dens = FALSE)
}

}
