% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stackedbar_facet.R
\name{plot_stackedbar_facet}
\alias{plot_stackedbar_facet}
\title{Plot a Stacked Bar Chart with Faceting and Labels}
\usage{
plot_stackedbar_facet(
  data,
  x,
  y,
  fill,
  facet,
  palette_name = "Paired",
  decimal = 0
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{x}{The variable to use for the x-axis (unquoted).}

\item{y}{The variable to use for the y-axis (unquoted, numeric).}

\item{fill}{The variable to use for the fill aesthetic (unquoted, categorical).}

\item{facet}{The variable to use for faceting (unquoted, categorical).}

\item{palette_name}{A string specifying the color palette from \code{RColorBrewer}. Defaults to \code{"Paired"}.}

\item{decimal}{An integer specifying the number of decimal places to round the group sum labels. Defaults to \code{0}.}
}
\value{
A ggplot object representing the stacked bar chart with optional faceting and labels.
}
\description{
Generates a stacked bar chart with required faceting. The function also includes labels
for percentages within each bar and total group sums above each bar. The y-axis is
formatted with commas, and bar colors can be customized using a specified color palette.
}
\examples{
\dontrun{
# Example dataset
library(dplyr)
df <- data.frame(
  year = rep(2021:2023, each = 6),
  category = rep(c("A", "B", "C", "D", "E", "F"), 3),
  value = c(100, 200, 150, 300, 250, 100, 200, 300, 400, 250, 150, 100, 300, 100, 200, 150, 400, 300),
  group = rep(c("G1", "G2"), 9)
)

# Generate a stacked bar chart
plot_stackedbar_facet(
  data = df,
  x = year,
  y = value,
  fill = category,
  facet = group,
  palette_name = "Set3",
  decimal = 1
)
}


}
