% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemap.R
\name{plot_treemap}
\alias{plot_treemap}
\title{Create a Treemap with Aggregated Subgroup Proportions}
\usage{
plot_treemap(data, type1, type2, value, type2_label = "Category")
}
\arguments{
\item{data}{A data frame containing the data for the treemap.}

\item{type1}{The primary grouping variable (unquoted).}

\item{type2}{The secondary grouping variable (unquoted).}

\item{value}{The numeric variable determining the area size (unquoted).}

\item{type2_label}{A character string specifying the legend title for \code{type2}. Defaults to "Category".}
}
\value{
A ggplot object representing the treemap.
}
\description{
This function generates a treemap visualization based on the input data. The treemap
groups data by a primary category (\code{type1}) and displays subgroups (\code{type2}) with sizes
proportional to their values (\code{value}). The primary group's proportions are labeled
as percentages.
}
\details{
This function uses the {ggplot2} and {treemapify} packages to generate
treemaps. Primary group percentages are calculated and displayed within their
respective tiles.
}
\examples{
\dontrun{
library(tidyverse)
library(treemapify)
library(see)

# Example data

# Generate the treemap
plot_treemap(df, type1, type2, value, type2_label = "Subcategories")
}

}
